Madness System 초벌구이

+++ Assets/Scripts/Components/NoneAvatarStatusComponent.cs
@@ -5,7 +5,7 @@ using Unity.Entities;
 
 [Serializable]
 public struct NoneAvatarStatusComponent : IComponentData {
-    public int madness;
+    public float madness;
 
     public NoneAvatarStatusComponent(ref NoneAvatarStatusComponent rhs) {
         madness = rhs.madness;
